+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 -Simple Stack ?.??.? (2019-01-14+)
 --------------------------------
 
+- FIX: Fix that a service registered multiple times in the same scope with different tags would receive service lifecycle callbacks as many times as it was registered.
+
 - ADD: Convenience method `stateChange.isTopNewStateEqualToPrevious()` to replace `stateChange.topNewState<Any>() == stateChange.topPreviousState()` condition check. It does the same thing, but maybe it's a bit easier to read.
 
 - FIX: Fix a typo which resulted in not throwing if the provided service tag was `null` (whoops. -_-)